+++ src/pipeline/hyper_heuristics/random.py
@@ -15,7 +15,7 @@ class RandomHyperHeuristic:
         self.iterations_scale_factor = iterations_scale_factor
 
     def run(self, env:BaseEnv) -> bool:
-        max_steps = env.construction_steps * self.iterations_scale_factor
+        max_steps = int(env.construction_steps * self.iterations_scale_factor)
         current_steps = 0
         while current_steps <= max_steps and env.continue_run:
             heuristic = random.choice(self.heuristic_pools)
